Move txMiningView to the mining package.
This moves txMiningView and txDescGraph to the mining package. This
reduces the coupling between the mempool and mining packages and allows
mining tests to use the actual txMiningView and txDescGraph
implementations.
